# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Andi Language Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Word {
  id: number;
  andiWord: string;
  /** @nullable */
  lemma?: string | null;
  russian: string;
  /** @nullable */
  english?: string | null;
  partOfSpeech: string;
  /** @nullable */
  nounClass?: string | null;
  /** @nullable */
  grammaticalFunction?: string | null;
  /** @nullable */
  root?: string | null;
  /** @nullable */
  affixes?: string | null;
  /** @nullable */
  morphology?: string | null;
  /** @nullable */
  phonetic?: string | null;
  /** @nullable */
  examples?: string | null;
  /** @nullable */
  dialect?: string | null;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  license?: string | null;
  /** @nullable */
  confidence?: number | null;
  /** @nullable */
  editorNotes?: string | null;
  /** @nullable */
  level?: string | null;
  createdAt: string;
}

export interface WordInput {
  andiWord: string;
  lemma?: string;
  russian: string;
  english?: string;
  partOfSpeech: string;
  nounClass?: string;
  grammaticalFunction?: string;
  root?: string;
  affixes?: string;
  morphology?: string;
  phonetic?: string;
  examples?: string;
  dialect?: string;
  source?: string;
  license?: string;
  confidence?: number;
  editorNotes?: string;
  level?: string;
}

export interface WordUpdate {
  andiWord?: string;
  lemma?: string;
  russian?: string;
  english?: string;
  partOfSpeech?: string;
  nounClass?: string;
  grammaticalFunction?: string;
  root?: string;
  affixes?: string;
  morphology?: string;
  phonetic?: string;
  examples?: string;
  dialect?: string;
  source?: string;
  license?: string;
  confidence?: number;
  editorNotes?: string;
  level?: string;
}

export interface WordList {
  words: Word[];
  total: number;
}

export interface WordForm {
  id: number;
  wordId: number;
  form: string;
  /** @nullable */
  caseName?: string | null;
  /** @nullable */
  caseNameRu?: string | null;
  /** @nullable */
  number?: string | null;
  /** @nullable */
  nounClass?: string | null;
  /** @nullable */
  grammarNote?: string | null;
}

export interface WordFormsResponse {
  word: Word;
  forms: WordForm[];
}

export interface AnalyzeWordInput {
  word: string;
}

export interface MorphSegment {
  text: string;
  type: string;
  label: string;
  labelRu?: string;
}

export interface MorphologicalAnalysis {
  input: string;
  isPreliminary: boolean;
  matchedWord?: Word | null;
  segments: MorphSegment[];
  explanation: string;
  /** @nullable */
  partOfSpeech?: string | null;
  /** @nullable */
  nounClass?: string | null;
  /** @nullable */
  case?: string | null;
  /** @nullable */
  number?: string | null;
  confidence?: number;
}

export interface Lesson {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  level: string;
  orderIndex: number;
  createdAt: string;
}

export interface Exercise {
  id: number;
  /** @nullable */
  lessonId?: number | null;
  type: string;
  prompt: string;
  /** @nullable */
  options?: string | null;
  answer: string;
  /** @nullable */
  explanation?: string | null;
  createdAt: string;
}

export interface LessonDetail {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  level: string;
  orderIndex: number;
  /** @nullable */
  content?: string | null;
  createdAt: string;
  exercises: Exercise[];
}

export interface LessonInput {
  title: string;
  description?: string;
  level: string;
  orderIndex: number;
  content?: string;
}

export interface LessonUpdate {
  title?: string;
  description?: string;
  level?: string;
  orderIndex?: number;
  content?: string;
}

export interface ExerciseInput {
  lessonId?: number;
  type: string;
  prompt: string;
  options?: string;
  answer: string;
  explanation?: string;
}

export interface ExerciseSubmission {
  answer: string;
}

export interface ExerciseResult {
  correct: boolean;
  correctAnswer: string;
  /** @nullable */
  explanation?: string | null;
}

export interface Flashcard {
  wordId: number;
  word: Word;
  dueDate: string;
  repetitions: number;
  easeFactor: number;
  interval: number;
}

export interface FlashcardReview {
  quality: number;
}

export interface ProgressSummary {
  wordsLearned: number;
  exercisesDone: number;
  streak: number;
  totalWords: number;
  lessonsCompleted: number;
  dueToday: number;
}

export interface CountEntry {
  label: string;
  count: number;
}

export interface StatsSummary {
  totalWords: number;
  totalLessons: number;
  totalExercises: number;
  partOfSpeechBreakdown: CountEntry[];
  levelBreakdown: CountEntry[];
}

export interface ActivityEntry {
  date: string;
  exercisesDone: number;
  wordsReviewed: number;
}

export interface CaseInfo {
  name: string;
  nameRu: string;
  suffix: string;
  function: string;
  example: string;
  exampleTranslation?: string;
}

export interface CaseParadigmsResponse {
  description: string;
  cases: CaseInfo[];
}

export interface NounClassInfo {
  classNumber: string;
  nameRu: string;
  marker: string;
  markerPosition?: string;
  semantics: string;
  examples: string[];
}

export interface NounClassesResponse {
  description: string;
  classes: NounClassInfo[];
}

export interface GrammarDrill {
  id: string;
  topic: string;
  topicRu?: string;
  question: string;
  answer: string;
  options?: string[];
  explanation: string;
}

export type ListWordsParams = {
search?: string;
partOfSpeech?: string;
nounClass?: string;
level?: string;
limit?: number;
offset?: number;
};

export type ListExercisesParams = {
lessonId?: number;
type?: string;
};

export type GetDueFlashcardsParams = {
limit?: number;
};

export type GetActivityFeedParams = {
days?: number;
};

export type GetGrammarDrillsParams = {
topic?: string;
};