# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Andi Language Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lessons:
    get:
      operationId: listLessons
      tags: [lessons]
      summary: List all lessons
      responses:
        "200":
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
    post:
      operationId: createLesson
      tags: [lessons]
      summary: Create a lesson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /lessons/{id}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get lesson by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonDetail"
        "404":
          description: Not found
    patch:
      operationId: updateLesson
      tags: [lessons]
      summary: Update a lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /exercises:
    get:
      operationId: listExercises
      tags: [exercises]
      summary: List exercises
      parameters:
        - name: lessonId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
    post:
      operationId: createExercise
      tags: [exercises]
      summary: Create an exercise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExerciseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
  /exercises/{id}/submit:
    post:
      operationId: submitExercise
      tags: [exercises]
      summary: Submit an answer to an exercise
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExerciseSubmission"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseResult"
  /flashcards/due:
    get:
      operationId: getDueFlashcards
      tags: [flashcards]
      summary: Get flashcards due for review
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Due flashcards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Flashcard"
  /flashcards/{wordId}/review:
    post:
      operationId: reviewFlashcard
      tags: [flashcards]
      summary: Record a flashcard review result
      parameters:
        - name: wordId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardReview"
      responses:
        "200":
          description: Updated flashcard state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flashcard"
...
components:
  schemas:
...
    Word:
      type: object
      required: [id, andiWord, russian, partOfSpeech, createdAt]
      properties:
        id:
          type: integer
        andiWord:
          type: string
        lemma:
          type: ["string", "null"]
        russian:
          type: string
        english:
          type: ["string", "null"]
        partOfSpeech:
          type: string
        nounClass:
          type: ["string", "null"]
        grammaticalFunction:
          type: ["string", "null"]
        root:
          type: ["string", "null"]
        affixes:
          type: ["string", "null"]
        morphology:
          type: ["string", "null"]
        phonetic:
          type: ["string", "null"]
        examples:
          type: ["string", "null"]
        dialect:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        license:
          type: ["string", "null"]
        confidence:
          type: ["number", "null"]
        editorNotes:
          type: ["string", "null"]
        level:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Lesson:
      type: object
      required: [id, title, level, orderIndex, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        level:
          type: string
        orderIndex:
          type: integer
        createdAt:
          type: string
    LessonDetail:
      type: object
      required: [id, title, level, orderIndex, createdAt, exercises]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        level:
          type: string
        orderIndex:
          type: integer
        content:
          type: ["string", "null"]
        createdAt:
          type: string
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/Exercise"
    LessonInput:
      type: object
      required: [title, level, orderIndex]
      properties:
        title:
          type: string
        description:
          type: string
        level:
          type: string
        orderIndex:
          type: integer
        content:
          type: string
    LessonUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        level:
          type: string
        orderIndex:
          type: integer
        content:
          type: string
    Exercise:
      type: object
      required: [id, type, prompt, answer, createdAt]
      properties:
        id:
          type: integer
        lessonId:
          type: ["integer", "null"]
        type:
          type: string
        prompt:
          type: string
        options:
          type: ["string", "null"]
        answer:
          type: string
        explanation:
          type: ["string", "null"]
        createdAt:
          type: string
    ExerciseInput:
      type: object
      required: [type, prompt, answer]
      properties:
        lessonId:
          type: integer
        type:
          type: string
        prompt:
          type: string
        options:
          type: string
        answer:
          type: string
        explanation:
          type: string
    ExerciseSubmission:
      type: object
      required: [answer]
      properties:
        answer:
          type: string
    ExerciseResult:
      type: object
      required: [correct, correctAnswer]
      properties:
        correct:
          type: boolean
        correctAnswer:
          type: string
        explanation:
          type: ["string", "null"]
    Flashcard:
      type: object
      required: [wordId, word, dueDate, repetitions, easeFactor, interval]
      properties:
        wordId:
          type: integer
        word:
          $ref: "#/components/schemas/Word"
        dueDate:
          type: string
        repetitions:
          type: integer
        easeFactor:
          type: number
        interval:
          type: integer
    FlashcardReview:
      type: object
      required: [quality]
      properties:
        quality:
          type: integer
...