# lib/api-zod/src/generated/types/lessonInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Andi Language Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface LessonInput {
  title: string;
  description?: string;
  level: string;
  orderIndex: number;
  content?: string;
}